Move shared code to HomogenousSedes
