+++ ssz/sedes/basic.py
@@ -167,6 +167,26 @@ class ProperCompositeSedes(BaseProperCompositeSedes[TSerializable, TDeserialized
 class HomogeneousProperCompositeSedes(
     ProperCompositeSedes[TSerializable, TDeserialized]
 ):
+    def __init__(self, element_sedes: TSedes, max_length: int) -> None:
+        self.element_sedes = element_sedes
+        if max_length < 1:
+            raise ValueError(
+                f"(Maximum) length of homogenous composites must be at least 1, got {max_length}"
+            )
+        self.max_length = max_length
+
     def get_sedes_id(self) -> str:
         sedes_name = self.__class__.__name__
         return f"{sedes_name}({self.element_sedes.get_sedes_id()},{self.max_length})"
+
+    @property
+    def is_packing(self) -> bool:
+        return isinstance(self.element_sedes, BasicSedes)
+
+    @property
+    def chunk_count(self) -> int:
+        if self.is_packing:
+            element_size = self.element_sedes.get_fixed_size()
+            return (element_size * self.max_length + CHUNK_SIZE - 1) // CHUNK_SIZE
+        else:
+            return self.max_length
